feat(handbook): Allow anonymous harassment reports
I wanted to add this section because it's an experience I've had in the
past when a friend worked in a toxic environment. Basically, you burn a
lot of capital when you make reports, and we concluded that the
behavior needed to stop, but that burning capital wasn't worth it. Our
solution was an anonymous report. It went well, but absent a clause like
this in my friend's employment manual, we were rather nervous about our
report being unmasked.

These kinds of reports are something that people can do whether you
endorse them or not. I prefer to endorse it and to set some expectations
as to how we will treat such reports.