feat(leave): Reduce pre-employment period to six months
I've waffled a bit on this, but I think six is better. When it's nine
months, you might take the job when you're one month pregnant and not
even know it. Then you're screwed. Since the gov't doesn't offer
parental leave and thus leaves it up to employers, we have to make
sure people get it.